Clear selection or autocomplete popup first before exiting to command mode (#5849)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->
Fixes #5844 .


https://github.com/user-attachments/assets/a0acf508-7f5e-4ce0-86ff-7f97cf82c1ea

This returns early when there is an autocomplete / text selection.
Subsequent escapes will exit to command mode.

## 🔍 Description of Changes

There might still be an issue of autocomplete appearing after `,` but is
not related to this i think.

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.