fix: handle double interruptions
This change makes the cell runner more robust to keyboard interrupts by
catching an additional KeyboardInterrupt when handling the original
KeyboardInterrupt. We've observed that interruptions are sometimes sent
twice, perhaps because some libraries forward signals in addition to our
own handler. In particular, without this change, interrupting
pl.read_parquet() would send two KeyboardInterrupts, killing the
runtime.

More work could be done to make the CellRunner more robust to
KeyboardInterrupts but this fixes the issue with polars in
local testing.