Allow deletion of nonlocal variables
This change makes it possible to delete variables that were defined in
another cell.

The motivation for this change is to allow advanced users to reclaim
memory. Especially when dealing with tensors or large datasets, either
reusing a variable name or explicitly deleting it with `del` are
traditionally used to reclaim memory. Reusing variables is not allowed
in marimo, and before this change `del` was not either, since it wasn't
clear how to order cells.

This change allows the del operator by making the following change to
dataflow graph construction:

  If a cell deletes a variable, it becomes a child of all other cells
  that reference that variable.

For example:

cell u
```
x
```

cell v
```
del x
```

In this case, v becomes a child of u.

This has the (desired) consequence that two cells cannot delete the same
variable, as that forms a cycle error.

This change does introduce some path dependency in how cells are
executed. In the above example, if cell u were manually re-run after
cell v ran, the user would get a NameError. We mitigate this by
communicating in the error message that the variable may have been
deleted (if at least one cell deletes the variable). While the path
dependency is unfortunate, using `del` is an advanced pattern that one
only uses when they're in pretty deep (and trying to avoid restarting
their kernel!), so the trading off execution reproducibility for
unblocking more resource intensive workflows is in my mind worth it.
Without this change, users would need to instead coalesce all
memory-intensive work into a single function (making incremental
computation difficult).