improvement: AI context registry + codemirror decorated resources (#5962)
This add a concept of an `AIContextProvider` and `AIContextRegistry`.
Providers can provide "context" to chat inputs. For example, context can
be variables or dataframes or sql tables (as we have today). This is
mostly a refactor to make adding new context easier.

This also brings in https://github.com/marimo-team/codemirror-mcp for
better `@` completions and plugin decorations with hover.