fix: stub and restore cache meta values  (#5996)
replaces #5933

---

## 📝 Summary

closes https://github.com/marimo-team/marimo/issues/5930

Process `self.meta["return"]` in the same way that `contextual_defs` are
handled in `cache.py` so that when caching and restoring UI elements
they are re-registered with the `UIElementRegistry`

## 🔍 Description of Changes

https://github.com/marimo-team/marimo/issues/5930 describes a bug from
`UIElement`s not being re-registered when they are returned from a
caching function.

It seemed like handling `UIElement`s (and some other classes) already
handled doing this correctly when it came to "contextual_defs" by using
a pattern of storing and loading "stub" classes, so I figured that using
that same process for the `return` value of the cached function should
work as well. It was either that or explicitly re-register `UIElement`s
on cache hit, which seemed like a worse solution.

I tried adding some tests for this but got kinda stuck on how to assert
that the old and new ui elements were "the same" since the stub loading
process technically creates a new instance. If we need tests for this, I
would appreciate some advice.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

---------

Co-authored-by: David Buckley <david.buckley@shopify.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>