improvement: use fully qualified model ids (#6003)
This changes model ids to be fully qualified `gpt-4o` ->
`openai/gpt-4o`. this handles old ids with a proper fallback, using the
same guessing logic to pick the right provider. It currently warns for
missing provider ids (in the frontend and server logs), but eventually
we can enforce this more.

This won't break existing config, but will show warnings in the UI and
terminal. This also will force you to change your model which won't be
backwards compatible