@marimo-team/llm-info package (#6005)
This PR adds a new package `@marimo-team/llm-info`. This is mainly to
keep things organized and make it easier for contributions. This is a
private package and not shared to npm (although could do that if people
wanted).

This does not add all the models or providers we use, but mainly a
placeholder to start to add more and let others jump in.

This PR also adds:
1. `@marimo-team/tsconfig` for shared config files
2. Adds icons in the settings page as an example:

<img width="727" height="247" alt="Screenshot 2025-08-11 at 2 40 27 PM"
src="https://github.com/user-attachments/assets/399b8259-9362-4494-b25b-253b7fafe349"
/>