fix: stdout and stderr should not be stoppable by users (#5955)
`sys.stdout` and `sys.stderr` are not stoppable by users in Python; our
mocked streams were stoppable (but they shouldn't be, not by user code),
which caused `Loguru:logger.remove()` to crash on Windows.

NB: I believe this fixes #3756, but I don't have a Windows machine to
test it on.

Makes progress on #3756