Proper handling of requests to /health, /.well-known and ./assets in the vite dev setup (#6135)
## 📝 Summary

Currently, when you run `pnpm dev` and open a notebook in a browser,
there are 2 requests that result in errors causing confusing log. Those
requests and the side effects are the following.

For the request:

```
http://127.0.0.1:2718/.well-known/appspecific/com.chrome.devtools.json
```

in the vite log we see this:

```
Failed to connect to a marimo server at http://127.0.0.1:2718/.well-known/appspecific/com.chrome.devtools.json

A marimo server may not be running.
Run marimo edit --no-token --headless in another terminal to start the server.

If the server is already running, make sure it is using port 2718 with --port=2718.
```

For the request:

```
http://127.0.0.1:2718/health
```

we see this:

```
Element title not found.
Element marimo-filename not found.
```

Also, there's a request from the browser to the asset hardcoded in
`_static/index.html` produces the following error in the browser
console:

```
(index):92  GET http://localhost:3000/assets/index-CVK4EV_D.js net::ERR_ABORTED 404 (Not Found)
```

## 🔍 Description of Changes

For the request to `.../.well-known...` we simply return 404.
For the request to `.../health` we return serverHtml directly without
processing.
We avoid the request to `./assets/...` by skipping copying this part to
the devHtml.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] Not sure what tests apply here.
- [x] I have run the code and verified that it works as expected.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>