improvement: store previous prompts in local storage. cycle through with up/down arrow (#6157)
This brings in the `promptHistory` plugin from `codemirror-ai` plugin to
allow up/down arrow to navigate the history.

There is a small edge-case since we share state across multiple prompts.
The state in one input can be stale if the input has been rendered and
we save the prompt in another input. But i think this is ok.

---------

Co-authored-by: Shahmir Varqha <Sham9871@gmail.com>