fix: remove unecessary work in scoped state (#6160)
This removes the usage of `jotai-scope` due to this bug
https://github.com/jotaijs/jotai-scope/issues/25

This would 1) cause unnecessary work (not terrible), and 2) re-execute
side-effectful atoms.

We probably should not have side-effectful atoms but one was the
runtime-manager which would check if it was healthy on creation. So each
new cell would trigger a `/healthy` check