fix: fixes for functions and anywidget virtual file when auto-instantiate=off (#6184)
Fixes #5866 

This fixes two issues that arise when `auto-instantiate=off`. 
1. The first is with virtual files. when a cell has not run, it has no
virtual files, but the old saved outputs try to request this, showing an
error. We still show the error, but running the cell again does fix the
issue.
2. The second is with function requests. when a cell has not run, it has
no registered function, but the old saved outputs try to request this,
showing an error. We still show the error, but running the cell again
does fix the issue.

In both cases, we show the error. This first PR lets the user fix
themselves by re-running the cell, which was not possible before. The
next step (in another PR) will be to either not make the request and
thus the error, or provide a clearer error that cell needs to be re-run