If the gpu stats command fails, don't run it again on every usage check (#6202)
## 📝 Summary

On a machine where the `nvidia-smi` command exists, but it's execution
fails for whatever reason (it cannot communicate with the driver, for
example), marimo's console logs are spammed with
`[W 250830 16:25:58 health:265] Failed to extract GPU stats: Command
'['nvidia-smi',
'--query-gpu=index,name,memory.total,memory.used,memory.free',
'--format=csv,noheader,nounits']' returned non-zero exit status 9`
This commit add a check further up the line, so that the warning is
displayed on the console only once.

## 🔍 Description of Changes
I added a simple invocation of `nvidia-smi` to an already existing
function which just used to check for the existence of the binary. If
the invocation fails for whatever reason, the warning to the console is
only printed once (because the result of the function is cached)

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] I have run the code and verified that it works as expected.