Implement timeout option (#6196)
## 📝 Summary

Add `--timeout` CLI flag which will cause the server to shutdown after
the specified duration if no connections are made or active. See #6163

## 🔍 Description of Changes

Added a Starlete Middleware, `TimeoutMiddleware`, which increments a
timer on the app state to record when the most recent activity was.
Every 10 seconds (currently not configurable), it compares to the
current time and shuts down the server if necessary.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable). #6163
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

---------

Co-authored-by: Myles Scolnick <myles@marimo.io>