feat: Reference files as context with # (#6241)
This adds file search with `#`. Instead of adding it to the context as a
resource, it attaches it as an attachment.
I could enable this for `@` as well, but currently the resources expect
static completions and not dynamic (like file search). This also does
not handle directories.

<img width="633" height="810" alt="Screenshot 2025-09-08 at 4 52 42 PM"
src="https://github.com/user-attachments/assets/0056b572-0dbb-4a99-95c9-9a50b7d319c4"
/>

This also:
* fixes reasoning parts that broke in the 5.0 ai sdk migration
* adds an `@` button to trigger context completion