Fallback to `__dict__`/`__slots___` in `enc_hook` for unknown objects (#6332)
Fixes #6330

Extends the `enc_hook` for msgspec with a fallback to `__dict__` /
`__slots__` for unknown types. Also adds back a `test_json_encoder.py`
which was removed becuase we deleted the `WebComponentEncoder`, but
adapted with expected outputs with `msgspec`.

A few observations:

- Overall, encoding "just works" however there are some edge cases that
were updated to msgspec-specific behavior, and it would be good to look
through the cases individually.

- `enc_hook` is only called if a type is unknown to `msgspec`.
Importantly, protocol-based serialization with `_mime_` only works if
the class is NOT a `msgspec.Struct` or `dataclasses.dataclass`

- As a result, you cannot override the encoding of any "known" types to
`msgspec` within the `enc_hook`. I had to update the timedelta test to
use [ISO 8901 duration
strings](https://en.wikipedia.org/wiki/ISO_8601#Durations) and
`float(inf | -inf | nan)` all serialize to `null`. To my knowledge there
is no way to intercept this. I added a test case just to make that
constraint explicit.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Myles Scolnick <myles@marimo.io>