fix: allow test cells in the presence of async cells (#6355)
## 📝 Summary

Enables async tests for in notebook auto-testing, e.g.

```python
async def my_test():
    assert True
```

Previously, this was only caught by pytest, not within the notebook
itself

## 🔍 Description of Changes

Instead of ignoring the type of the result of run for a test cell, and
assuming that it's synchronous, check for whether it's awaitable, and
await it if so.

## 📋 Checklist

- [ ] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [ ] I have run the code and verified that it works as expected.

---------

Co-authored-by: Sasha Lopoukhine <superlopuh@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>