add uneccesary use-effects eslint plugin and remove some use-effects (#6454)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->
package:
https://github.com/NickvanDyke/eslint-plugin-react-you-might-not-need-an-effect
reddit inspo:
https://www.reddit.com/r/reactjs/comments/1nm09bq/update_eslint_plugin_to_catch_unnecessary/
good read: https://react.dev/learn/you-might-not-need-an-effect#

There are other warnings, but these seemed the easiest to drive-by fix.
The plugin is not exhaustive, the one in datasources was largely not
caught. I think it's still useful to have as it catches many common
issues.

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.