Support sticky headers on table  (#6486)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->

Sorry if this is more of an feature request than an actual PR.
We have long and wide tables in our organization and wish to improve
scroll through this vertically.
(We are using `pagination=False` to see all rows)

`freeze_columns` works well nicely for horizontal scrolling.
But when a table is long:
- The table takes the entire page, and the scrollbar is on the page
level.
- The header is gone as soon as your scroll.

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

It would be interesting if we could have something long these lines:

![Kapture 2025-09-23 at 11 32
25](https://github.com/user-attachments/assets/578e7bbc-3924-49fc-9221-a993e4acb4d9)

- Scrollbar is on the table
- Header is sticky

(table header is red for dramatic effect)

Could we have something that brings this experience?
If so, I'd be happy to work on this in whatever solution you deem best.

//cc @mscolnick @Light2Dark 


## 📋 Checklist

- [ ] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [ ] I have added tests for the changes made.
- [ ] I have run the code and verified that it works as expected.