add parse function in /validate endpoint for SQL (#6517)
## 📝 Summary

<!--
Provide a concise summary of what this pull request is addressing.

If this PR fixes any issues, list them here by number (e.g., Fixes
#123).
-->

Adds a parse function in `/validate` endpoint that currently supports
duckdb so we get better lint errors when writing SQL cells.

/validate runs `SELECT JSON_SERIALIZE_SQL` in duckdb, which returns a
structured error response (for a majority of queries)
https://duckdb.org/docs/stable/data/json/sql_to_and_from_json.html.

We then convert the json response to linter error format
```python
    message: str
    line: int
    column: int
    severity: Literal["error", "warning"]
```

A custom sql parser is created to use this new endpoint, everytime the
parser runs `validateSql` for DuckDB, it will call this endpoint instead
of using NodeSqlParser default parsing.

This PR also disables the `EXPLAIN` mode because they use the same
endpoint and so there is a perf hit when called twice. We can optimize
the endpoint.

### Endpoint design:
`/sql/validate`

There are 2 possible paths:
- catalog check : Checks the query against the database by using
`EXPLAIN {query}`. Surfaces both syntax and catalog errors (invalid
table_name etc.) but is not structured.
- parse sql : Uses AST/backend parser to validate the query. This
usually does not catch all errors but is cheaper than validating against
a database. Only surfaces syntax errors, and not all syntax errors. Is
structured.

When onlyParse is true, we just run the parse function, which is the aim
of this PR. In the future, we want to optimize so that we can query once
and get the complete response back.

## 🔍 Description of Changes

<!--
Detail the specific changes made in this pull request. Explain the
problem addressed and how it was resolved. If applicable, provide before
and after comparisons, screenshots, or any relevant details to help
reviewers understand the changes easily.
-->

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.

---------

Co-authored-by: Myles Scolnick <myles@marimo.io>