don't override auto-save by when in --watch mode (#6535)
Fixes #6534

Closes https://github.com/marimo-team/marimo/issues/4194
Closes https://github.com/marimo-team/marimo/issues/4196

This updates the behavior of watch mode in the server configuration to
ensure that user-defined auto-save is kept, rather than being forcibly
overridden. We had this originally enabled to avoid conflicts between
the server's view of the file and external changes made from the
filesystem. Since the cell matching has improved
(https://github.com/marimo-team/marimo/pull/6127), I don't expect this
to cause issues.

We can in a followup handle overwrites and show the user (similar to how
vscode or other IDEs might handle this)

---------

Co-authored-by: Trevor Manz <trevor.j.manz@gmail.com>