feat: edit files from a Gist url (#6559)
## 📝 Summary

Adds support for editing files from a Github Gist url. Closes #5950.

## 🔍 Description of Changes

There are two ways to enter a Gist url:

1. Direct Raw URLs
(`gist.github.com/user/gist_id/raw/file_id/filename`): We fetch them
directly.
2. Standard Gist URLs (`gist.github.com/user/gist_id`): For these, we
make a request to the GitHub Gist API (`api.github.com/gists/gist_id`)
to retrieve metadata. Then we parse the response to find the raw URL of
the first Python or Markdown file in the gist.

The code is organized like the Github url flow, with the exception that
the extension check is not in the `can_read` function but in the `read`
function.

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [ ] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions) (Please
provide a link if applicable).
- [x] I have added tests for the changes made.
- [x] I have run the code and verified that it works as expected.