Click filenames in stack trace to open locally in editor (#6556)
## 📝 Summary

Fixes #6390

## 🔍 Description of Changes

* When parsing tracebacks, in addition to marimo cells replacing the
traceback with a link and button, now other files are also wrapped with
a clickable div element, which invokes the user's `EDITOR` to open the
file
* `marimo._server.files.OSFileSystem.open_in_editor` now supports
opening a file in a specific line, which the traceback uses to open the
relevant line from the trace

## 📋 Checklist

- [X] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [X] I have added tests for the changes made
- I maintained the existing test for processing traceback HTMLs to work
with the new changes
- [X] I have run the code and verified that it works as expected.

## Btw (devops)

I tried to follow the instructions according to `CONTRIBUTING.md`, but:

* I had to uninstall an older local version of node for things to work,
otherwise node resolved to the oldest version I have which was too old?
* Couldn't get `pnpm` to work so had to use `make fe` for an >1min dev
feedback cycle. (btw I'm a front-end noob so I might be missing basic
things)