refactor: MPL proxy endpoint (#6632)
## 📝 Summary
This PR addresses an issue in the `/mpl/<port>/<route>` endpoint that
allowed unauthenticated access to internal services on any port.

## Changes Made
- Removed the MPL middleware proxy endpoint that indiscriminately
forwarded ports
 - Replaced associated proxy routing logic as part of main server
 - Validate proxy request with figure id

See previous behavior issue in tests/_cli/test_endpoint.py

## Credits
Research and disclosure by @devgi and @acepace.

---
Played with a couple ways of doing this, but I think this is the best. I
was unable to get any surface area when playing around. Added a unit
test that fails in the exploitable case