docs: change READMEs to be more consistent for different languages (#6696)
## 📝 Summary

Edit all of the README files to be more consistent across languages.

## 🔍 Description of Changes

1. Make indentation/spacing more consistent
2. Add missing `</p>` tag
3. Make Spanish README list the languages in their respective spelling
(instead of Spanish for all of them). This is consistent with the other
README files
4. Break up "nav-bar" HTML so that the `<b> | </b>` separators are
between every language
- This makes it easier to keep the README files in sync, because if you
switch between them, the only difference is which languages are wrapped
by an `<a>` tag (also helped by change 3)
5. Fix outdated links in Chinese and Spanish README files

## 📋 Checklist

- [x] I have read the [contributor
guidelines](https://github.com/marimo-team/marimo/blob/main/CONTRIBUTING.md).
- [x] For large changes, or changes that affect the public API: this
change was discussed or approved through an issue, on
[Discord](https://marimo.io/discord?ref=pr), or the community
[discussions](https://github.com/marimo-team/marimo/discussions).
    - N/A
- [x] I have added tests for the changes made.
    - N/A
- [x] I have run the code and verified that it works as expected.
    - N/A